Function which checks if a certain team wins. Check Health Function. TurnBased UI
Functions that check if the health of each character is zero, if it is
then destroy that actor.

There are two functions which check if all human or AI players are dead
which are to be called when damage is dealt.

Combat UI has been added with a move and attack button instead of
pressing 1 and 2